14/02

+++ OO2/src/ej47/Articulo.java
@@ -1,16 +1,18 @@
 package ej47;
 
+import java.math.BigDecimal;
+
 public abstract class Articulo {
 
-	protected Integer precio;
+	protected BigDecimal precio;
 	protected String descripcion;
 	
-	public Articulo(Integer precio, String descripcion) {
+	public Articulo(BigDecimal precio, String descripcion) {
 		this.precio = precio;
 		this.descripcion = descripcion;
 	}
 
-	public Integer getPrecio() {
+	public BigDecimal getPrecio() {
 		return precio;
 	}
